Change uint16_t to uint64_t for linker

+++ src/front/asm/generate.cpp
@@ -791,7 +791,7 @@ int generate(const vector<string>& expanded_lines, const map<long unsigned, long
     /////////////////////////////////////////////////////////
     // GATHER LINKS, GET THEIR SIZES AND ADJUST BYTECODE SIZE
     vector<string> links = assembler::ce::getlinks(ilines);
-    vector<tuple<string, uint16_t, char*> > linked_libs_bytecode;
+    vector<tuple<string, uint64_t, char*> > linked_libs_bytecode;
     vector<string> linked_function_names;
     vector<string> linked_block_names;
     map<string, vector<unsigned> > linked_libs_jumptables;
@@ -831,7 +831,7 @@ int generate(const vector<string>& expanded_lines, const map<long unsigned, long
             }
         }
 
-        linked_libs_bytecode.push_back( tuple<string, uint16_t, char*>(lnk, loader.getBytecodeSize(), loader.getBytecode()) );
+        linked_libs_bytecode.push_back( tuple<string, uint64_t, char*>(lnk, loader.getBytecodeSize(), loader.getBytecode()) );
         bytes += loader.getBytecodeSize();
     }
